+++ src/trans.cpp
@@ -89,7 +89,7 @@ mrb_value siren_trans_to_s(mrb_state* mrb, mrb_value self)
       "    Y = % 2.6f, % 2.6f, % 2.6f\n"
       "    Z = % 2.6f, % 2.6f, % 2.6f\n"
       "    T = % 2.6f, % 2.6f, % 2.6f>",
-      (uintptr_t)mrb_cptr(self),
+      (unsigned int)(uintptr_t)mrb_cptr(self),
       a11, a21, a31,
       a12, a22, a32,
       a13, a23, a33,
